# server/src/credits/dto/transaction-history-response.dto.ts
import { TransactionDto } from './transaction.dto';

/**
 * Transaction history response DTO
 */
export class TransactionHistoryResponseDto {
  transactions: TransactionDto[];
  meta: {
    total: number;
    limit: number;
    offset: number;
  };
}